+++ src/game/skull_game/util.rs
@@ -5,8 +5,8 @@ use opencv::imgcodecs::{ImreadModes, imread};
 use opencv::imgproc::{COLOR_BGR2GRAY, cvt_color};
 use opencv::{Result, prelude::*};
 
-pub fn generate_index_for_quad(counter: usize, index_buffer_data: &mut Vec<u16>) {
-    let num = counter as u16;
+pub fn generate_index_for_quad(counter: usize, index_buffer_data: &mut Vec<u32>) {
+    let num = counter as u32;
     index_buffer_data.push(num * 4);
     index_buffer_data.push(num * 4 + 1);
     index_buffer_data.push(num * 4 + 2);
